requirements.txt
matplotlib
Pillow
numpy
tensorflow
streamlit